Use Crypto insted of hashlib, move function to utils.py
